+++ CHANGES.md
@@ -5,6 +5,7 @@
 * Methods on `AcceptLanguage`: `ranked` and `preference`. [#2480]
 * `AcceptEncoding` typed header. [#2482]
 * `Range` typed header. [#2485]
+* `ContentLength` typed header. [#????]
 * `HttpResponse::map_into_{left,right}_body` and `HttpResponse::map_into_boxed_body`. [#2468]
 * `ServiceResponse::map_into_{left,right}_body` and `HttpResponse::map_into_boxed_body`. [#2468]
 
@@ -25,6 +26,7 @@
 [#2482]: https://github.com/actix/actix-web/pull/2482
 [#2484]: https://github.com/actix/actix-web/pull/2484
 [#2485]: https://github.com/actix/actix-web/pull/2485
+[#????]: https://github.com/actix/actix-web/pull/????
 
 
 ## 4.0.0-beta.13 - 2021-11-30